R3_input
The provided Java code demonstrates how to use the `Scanner` class to take user input from the console. The program prompts the user to enter a name by displaying "Enter:". It then reads the input using `sc.next()` and stores it in the variable `name`. Finally, it prints out the entered name with the message "Name is " followed by the input provided by the user. This basic example illustrates the use of `Scanner` for reading input and concatenating strings for output.